api/requirements.txt
torch==2.6.0
transformers==4.48.3
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
accelerate==0.26.0
safetensors==0.5.2
python-dotenv==1.0.1
openai==1.63.0